2686: Added `access_token_expiry` field to `userSession` table to allow for web-config-server to know when its token is out of date
