+++ packages/server-boilerplate/package.json
@@ -34,7 +34,6 @@
     "body-parser": "^1.18.3",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
-    "express": "^4.16.2",
-    "jsonwebtoken": "^8.5.1"
+    "express": "^4.16.2"
   }
 }
